Code/b2b_backend/requirements.txt
Django>=5.0
djangorestframework
django-cors-headers
firebase-admin
django-environ
gunicorn
